requirements.txt
# AegisAI - Smart City Risk Intelligence System
# Python 3.10+ required

# Phase 1: Perception Layer
ultralytics>=8.0.0
deep-sort-realtime>=1.3.2
opencv-python>=4.8.0
numpy>=1.24.0

# Phase 4: Response Layer (API)
fastapi>=0.104.0
uvicorn>=0.24.0

# Database (PostgreSQL)
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
alembic>=1.13.0

# Caching (Redis)
redis>=5.0.0

# AI/LLM (Google Gemini via HTTP)
requests>=2.31.0

# Sprint 1: Security & Testing
python-dotenv>=1.0.0
slowapi>=0.1.9
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0

# Optional: GPU Acceleration (uncomment if using CUDA)
# torch>=2.0.0
# torchvision>=0.15.0
